Extracted OppositeSelectionDialog into own file
